+++ src/main/java/mod/emt/thaumictweaker/commands/CommandBaseTT.java
@@ -61,7 +61,9 @@ public class CommandBaseTT extends CommandBase {
 
     @Override
     public void execute(@NotNull MinecraftServer server, @NotNull ICommandSender sender, String[] args) throws CommandException {
-        if(args.length >= 1) {
+        if(args.length == 1) {
+            throw new CommandException(new TextComponentTranslation("command.thaumictweaker:invalid").getFormattedText());
+        }else if(args.length >= 1) {
             if(!subCommands.containsKey(args[0])) {
                 throw new CommandException(new TextComponentTranslation("command.thaumictweaker:invalid").getFormattedText());
             } else {